@@ -10,12 +10,12 @@ autoVersion {
 }
 
 android {
-	compileSdkVersion 27
+	compileSdkVersion 28
 
 	defaultConfig {
 		applicationId "pw.janyo.whatanime"
 		minSdkVersion 19
-		targetSdkVersion 27
+		targetSdkVersion 28
 		versionCode autoVersion.code
 		versionName autoVersion.name
 		setProperty("archivesBaseName", "WhatAnime-$versionName")
@@ -34,27 +34,26 @@ android {
 }
 
 dependencies {
-	implementation fileTree(include: ['*.jar'], dir: 'libs')
-	implementation 'com.android.support:support-v4:27.1.0'
-	implementation 'com.android.support.constraint:constraint-layout:1.0.2'
-	implementation 'com.google.firebase:firebase-core:12.0.0'
-	implementation 'com.android.support:appcompat-v7:27.1.0'
-	implementation 'com.android.support:design:27.1.0'
-	implementation 'com.android.support:cardview-v7:27.1.0'
+	implementation 'androidx.legacy:legacy-support-v4:1.0.0-beta01'
+	implementation 'androidx.constraintlayout:constraintlayout:2.0.0-alpha1'
+	implementation 'com.google.firebase:firebase-core:16.0.1'
+	implementation 'androidx.appcompat:appcompat:1.0.0-beta01'
+	implementation 'com.google.android.material:material:1.0.0-beta01'
+	implementation 'androidx.cardview:cardview:1.0.0-beta01'
 	implementation 'io.reactivex.rxjava2:rxjava:2.1.10'
 	implementation 'io.reactivex.rxjava2:rxandroid:2.0.2'
 	implementation 'com.github.bumptech.glide:glide:4.6.1'
 	annotationProcessor 'com.github.bumptech.glide:compiler:4.6.1'
-	implementation 'androidx.core:core-ktx:0.2'
-	implementation 'com.github.Mystery0Tools:Logs:1.0.0'
-	implementation 'com.github.Mystery0Tools:Tools:1.0.1'
+	implementation 'androidx.core:core-ktx:1.0.0-beta01'
+	implementation 'com.github.Mystery0Tools:Logs:1.2.0'
+	implementation 'com.github.Mystery0Tools:Tools:1.4.0'
 	implementation 'com.squareup.retrofit2:retrofit:2.4.0'
 	implementation 'com.squareup.retrofit2:adapter-rxjava2:2.4.0'
 	implementation 'com.squareup.retrofit2:converter-gson:2.4.0'
 	implementation 'com.squareup.okhttp3:okhttp:3.10.0'
 	implementation 'com.getkeepsafe.taptargetview:taptargetview:1.11.0'
-	implementation 'org.litepal.android:core:1.6.1'
-	implementation 'com.zyao89:zloading:1.0.11'
+	implementation 'org.litepal.android:core:2.0.0'
+	implementation 'com.zyao89:zloading:1.1.2'
 	implementation('com.crashlytics.sdk.android:crashlytics:2.7.1@aar') {
 		transitive = true
 	}
