@@ -2,9 +2,9 @@ package pw.janyo.whatanime.util.whatanime;
 
 import android.content.Context;
 import android.os.Environment;
-import android.support.design.widget.CoordinatorLayout;
-import android.support.design.widget.Snackbar;
-import android.support.v4.content.ContextCompat;
+import androidx.coordinatorlayout.widget.CoordinatorLayout;
+import com.google.android.material.snackbar.Snackbar;
+import androidx.core.content.ContextCompat;
 
 import com.google.gson.Gson;
 import com.zyao89.view.zloading.ZLoadingDialog;
